API:_Add ar_entry_get_raw_name function
Instead of using a zip-specific function for getting raw names,
which could be abused by calling it on non-zip archives, the api should
be extenced to allow getting a raw filename from all types of archives
if they require it.

To do so, the internal prototype for getting filenames is extended using
a "raw" parameter to indicate if the raw filename is requested and an
api function for requesting the raw name is added.

As raw filenames are only needed for malformed zip archives at this point,
all other types of archives will return NULL.