Fix heap corruption caused by unsigned integer underflow in rar rgb and e8 filter
Due to the usage of unsigned integer subtraction in some of the rar e8 and rgb
for loop conditions, an integer underflow could be provoked by supplying a
blocklength smaller than the calculations subtrahend.

Solution: Properly check if the blocklength is sufficient.

Note: This issue was found by Radoslaw Madej from Check Point Research